✨ feat: Add configuration options for replicas and max concurrent reconciles
This commit adds two new configuration options, "replicas" and "max concurrent reconciles", to the stable/integration-operator/questions.yaml file. These options allow users to specify the number of replicas and the maximum number of concurrent reconciles for the operator.

The default value for "replicas" is 1, and the default value for "max concurrent reconciles" is 3. These values can be overridden by the user in the stable/integration-operator/values.yaml file.

Additionally, this commit updates the stable/integration-operator/templates/deployments/integration-operator.yaml file to include the new configuration options. The number of replicas is now set using the "replicas" field, and the "--leader-elect" and "--health-probe-bind-address=:8081" arguments are added to the container command. Finally, liveness and readiness probes are added to the container to monitor its health.