✨ feat: Add patch-operator
The commit adds the `patch-operator` directory containing relevant files to the repository. It includes a `.helmignore` file with several ignored paths, a `Makefile` with a prepack target and a dependency on `mkpm.mk`, and a `values.yaml` file for configuring resources. The `OWNERS` file has also been deleted from the `patch-operator` directory.