feat(patch-operator): Add serviceAccountName field to PatchSpec
This commit adds the `serviceAccountName` field to the `PatchSpec` in the `patch.rock8s.com_patches.yaml` file located at `beta/patch-operator/templates/crds`.

The `serviceAccountName` field is described as the service account name used in the job.

The commit also includes other changes to the `spec` and `status` sections of the `Patch` resource.