✨ feat: Update _helpers.tpl for beta/aws-creds and beta/s3
Updated the _helpers.tpl file in the beta/aws-creds and beta/s3 directories. The changes include adding an "else" condition that outputs "local" when the clusterInfo data is not available. Additionally, modified the "s3.bucket-name" definition to concatenate .Values.bucket.name with the result of including "s3.cluster-name", replacing "." with "-".