♻️ refactor: Remove unused code and update replication configuration
This commit removes the unused file `ca-replicator.yaml` from the `integrations` directory. Additionally, it adds a new patch file `phpldapadmin.yaml` to the `patches` directory. The patch file enables Istio and modifies the `PHPLDAPADMIN_LDAP_HOSTS` configuration value. Lastly, in the `openldap.yaml` release file, the replication configuration is updated to use the `gt` function for comparing replicas and ensures the value is converted to an integer.