fix(openldap): Fix template for socket.yaml
This commit fixes the template for the `socket.yaml` file in the `openldap` directory. The changes include updating the values for the `tlsCrt` and `caCrt` variables. The `tlsCrt` variable now uses a ternary operator to handle cases where `BASE64_ENCODED_TLS_CRT` is not present, and the `caCrt` variable follows a similar pattern. This ensures that the appropriate decoding is performed when the variables are present.

The specific changes made are as follows:
- Replaced the expression `'{% .vars.BASE64_ENCODED_TLS_CRT | ternary (.vars.BASE64_ENCODED_TLS_CRT | b64dec) "" %}'` with `' {% (not .vars.BASE64_ENCODED_TLS_CRT) | "" ternary (.vars.BASE64_ENCODED_TLS_CRT | b64dec) %}'` for the `tlsCrt` variable.
- Replaced the expression `'{% .vars.BASE64_ENCODED_CA_CRT | ternary (.vars.BASE64_ENCODED_CA_CRT | b64dec) "" %}'` with `' {% (not .vars.BASE64_ENCODED_CA_CRT) | "" ternary (.vars.BASE64_ENCODED_CA_CRT | b64dec) %}'` for the `caCrt` variable.

These changes ensure that the `tlsCrt` and `caCrt` variables are handled correctly based on whether their corresponding values are present or not.