♻️ refactor: Update commonName in openldap.yaml template
The commonName field in the openldap.yaml template has been updated to include the service name and namespace. The previous commonName value was {{$serviceName | quote}}, and it has been changed to {{(printf "%s.%s.svc.cluster.local" $serviceName .Release.Namespace) | quote}}. This update ensures that the generated certificates have the correct commonName for the OpenLDAP service.