✨ feat: Update keycloak.yaml and keycloak-ingress.yaml templates
This commit updates the `keycloak.yaml` and `keycloak-ingress.yaml` templates in the `beta/keycloak/templates` directory.

In `keycloak-ingress.yaml`, the `port.number` field has been modified to use a conditional expression based on the values of `.Values.ingress.keycloak.enabled` and `.Values.ingress.keycloak.tls`. The updated value is now dynamically set to either "8443" or "8080".

In `keycloak.yaml`, a new command has been added to import a certificate using `keytool`. The command imports the specified certificate file into the keystore with the corresponding alias and password. This ensures that the certificate is properly configured.

Additionally, the `exec /opt/keycloak/bin/kc.sh start` command in the `spec.containers` section of the `keycloak.yaml` template has been modified to include the `--optimized` flag. This flag enables optimized startup behavior for Keycloak.

These changes enhance the functionality and security of the Keycloak deployment.