✨ feat: Add rbac.yaml and rename role.yaml to rbac.yaml
This commit adds a new file `rbac.yaml` under the `beta/keycloak/templates` directory. It also renames the file `role.yaml` to `rbac.yaml` under the `beta/postgres/templates` directory. The changes include defining a `RoleBinding` and a `Role` with associated metadata and labels. The `RoleBinding` binds a `Role` to a service account named "default" in the specified namespace. The `Role` has rules allowing unrestricted access to all resources and API groups.